Move getFormattedPersonInformationSelector PeopleSelector->PersonSelector
[#152463649]
